Added implementation for end to end encryption using the Station to Station protocol.
RsaKeyPair class contains static functions to create
cryptographic RSA Diffe Hellam (DH) and AES keys.

It also contain static database acessor methods to save
and retreive these keys into and from databases when
they are created or needed when a message is being sent.

The session between two parties, Initiator and Recipient
begins with Initiator sending a friend request to Recipient

The session is established in this order

Initiator
   ContactsActivity : InitiatorFriendRequest

Recipient:
		NetworkService.onFriendRequest

Initiator:

		NetworkService.OnFriendRequestReply

Note that as of now, the final part of the protocol
where Initiator and Recipient verify each other's
digital signatures has not yet been implemented. but
will be part of a future update.